fix(data-warehouse): Normalize primary key (#28940)
